wip: Vehicle drive setup to PVD
