+++ DustRiders/DustRiders/DustRiders.vcxproj
@@ -189,6 +189,7 @@
     <ClCompile Include="snippetvehicle2directdrive\SnippetVehicleDirectDrive.cpp" />
     <ClCompile Include="snippetvehicle2directdrive\SnippetVehicleRender.cpp" />
     <ClCompile Include="Texture.cpp" />
+    <ClCompile Include="Vehicle.cpp" />
     <ClCompile Include="VertexArray.cpp" />
     <ClCompile Include="VertexBuffer.cpp" />
     <ClCompile Include="Window.cpp" />
@@ -225,6 +226,7 @@
     <ClInclude Include="snippetvehicle2common\SnippetVehicleHelpers.h" />
     <ClInclude Include="Texture.h" />
     <ClInclude Include="Transform.h" />
+    <ClInclude Include="Vehicle.h" />
     <ClInclude Include="VertexArray.h" />
     <ClInclude Include="VertexBuffer.h" />
     <ClInclude Include="Window.h" />
