+++ DustRiders/DustRiders/PhysicsSystem.cpp
@@ -1,5 +1,9 @@
 #include "PhysicsSystem.h"
 
+using namespace physx;
+using namespace physx::vehicle2;
+using namespace snippetvehicle2;
+
 PhysicsSystem::PhysicsSystem()
 {
 	// Initialize PhysX
@@ -40,39 +44,49 @@ PhysicsSystem::PhysicsSystem()
 
 	// Simulate
 	gMaterial = gPhysics->createMaterial(0.5f, 0.5f, 0.6f);
-	physx::PxRigidStatic* groundPlane = physx::PxCreatePlane(*gPhysics, physx::PxPlane(0, 1, 0, 50), *gMaterial);
+	groundPlane = physx::PxCreatePlane(*gPhysics, physx::PxPlane(0, 1, 0, 0.01), *gMaterial);
+	for (PxU32 i = 0; i < groundPlane->getNbShapes(); i++)
+	{
+		PxShape* shape = NULL;
+		groundPlane->getShapes(&shape, 1, i);
+		shape->setFlag(PxShapeFlag::eSCENE_QUERY_SHAPE, true);
+		shape->setFlag(PxShapeFlag::eSIMULATION_SHAPE, false);
+		shape->setFlag(PxShapeFlag::eTRIGGER_SHAPE, false);
+	}
 	gScene->addActor(*groundPlane);
 
-	// Define a box
-	float halfLen = 0.5f;
-	physx::PxShape* shape = gPhysics->createShape(physx::PxBoxGeometry(halfLen, halfLen, halfLen), *gMaterial);
-	physx::PxU32 size = 30;
-	physx::PxTransform tran(physx::PxVec3(0));
+	PxInitVehicleExtension(*gFoundation);
 
-	// Save some space for all the boxes in the scene
-	rigidDynamicList.reserve(465);
-	transformList.reserve(465);
+	//// Define a box
+	//float halfLen = 0.5f;
+	//physx::PxShape* shape = gPhysics->createShape(physx::PxBoxGeometry(halfLen, halfLen, halfLen), *gMaterial);
+	//physx::PxU32 size = 30;
+	//physx::PxTransform tran(physx::PxVec3(0));
 
-	// Create a pyramid of physics-enabled boxes
-	for (physx::PxU32 i = 0; i < size; i++)
-	{
-		for (physx::PxU32 j = 0; j < size - i; j++)
-		{
-			physx::PxTransform localTran(physx::PxVec3(physx::PxReal(j * 2) - physx::PxReal(size - i), physx::PxReal(i * 2 - 1), 0) * halfLen);
-			physx::PxRigidDynamic* body = gPhysics->createRigidDynamic(tran.transform(localTran));
+	//// Save some space for all the boxes in the scene
+	//rigidDynamicList.reserve(465);
+	//transformList.reserve(465);
 
-			// Add the PxRigidDynamic to a vector
-			rigidDynamicList.push_back(body);
-			transformList.emplace_back(new Transform());
+	//// Create a pyramid of physics-enabled boxes
+	//for (physx::PxU32 i = 0; i < size; i++)
+	//{
+	//	for (physx::PxU32 j = 0; j < size - i; j++)
+	//	{
+	//		physx::PxTransform localTran(physx::PxVec3(physx::PxReal(j * 2) - physx::PxReal(size - i), physx::PxReal(i * 2 - 1), 0) * halfLen);
+	//		physx::PxRigidDynamic* body = gPhysics->createRigidDynamic(tran.transform(localTran));
 
-			body->attachShape(*shape);
-			physx::PxRigidBodyExt::updateMassAndInertia(*body, 10.0f);
-			gScene->addActor(*body);
-		}
-	}
+	//		// Add the PxRigidDynamic to a vector
+	//		rigidDynamicList.push_back(body);
+	//		transformList.emplace_back(new Transform());
 
-	// Clean up
-	shape->release();
+	//		body->attachShape(*shape);
+	//		physx::PxRigidBodyExt::updateMassAndInertia(*body, 10.0f);
+	//		gScene->addActor(*body);
+	//	}
+	//}
+
+	//// Clean up
+	//shape->release();
 
 	// Initialize transforms;
 	updateTransforms();
@@ -86,12 +100,22 @@ PhysicsSystem::PhysicsSystem()
 //	return position;
 //}
 
-void PhysicsSystem::updatePhysics()
+long i = 0;
+
+void PhysicsSystem::updatePhysics(double t)
 {
-	gScene->simulate(1.f / 60.f);
+	if (i < 15) { // First few renders should be simulated with manual step to avoid objects clipping through ground
+		gScene->simulate(1.f / 60.f);
+		i++;
+	}
+	else
+		gScene->simulate((t - lastTime) / 1.f);
 	gScene->fetchResults(true);
 
 	this->updateTransforms();
+	lastTime = t;
+
+	
 }
 
 void PhysicsSystem::updateTransforms()
