+++ DustRiders/DustRiders/PhysicsSystem.h
@@ -4,6 +4,9 @@
 #include <vector>
 
 #include "PxPhysicsAPI.h"
+#include "vehicle2/PxVehicleAPI.h"
+#include "snippetvehicle2common/serialization/BaseSerialization.h"
+
 #include "Transform.h"
 
 class PhysicsSystem
@@ -21,10 +24,14 @@ public:
 	physx::PxScene* gScene = NULL;
 	physx::PxMaterial* gMaterial = NULL;
 	physx::PxPvd* gPvd = NULL;
+	physx::PxRigidStatic* groundPlane = NULL;
 
 	PhysicsSystem();
 
 	//physx::PxVec3 getPosition();
-	void updatePhysics();
+	void updatePhysics(double);
 	void updateTransforms();
+
+private:
+	double lastTime = 0.0f;
 };
