+++ DustRiders/DustRiders/snippetvehicle2directdrive/SnippetVehicleRender.cpp
@@ -1,90 +1,90 @@
-// Redistribution and use in source and binary forms, with or without
-// modification, are permitted provided that the following conditions
-// are met:
-//  * Redistributions of source code must retain the above copyright
-//    notice, this list of conditions and the following disclaimer.
-//  * Redistributions in binary form must reproduce the above copyright
-//    notice, this list of conditions and the following disclaimer in the
-//    documentation and/or other materials provided with the distribution.
-//  * Neither the name of NVIDIA CORPORATION nor the names of its
-//    contributors may be used to endorse or promote products derived
-//    from this software without specific prior written permission.
+//// Redistribution and use in source and binary forms, with or without
+//// modification, are permitted provided that the following conditions
+//// are met:
+////  * Redistributions of source code must retain the above copyright
+////    notice, this list of conditions and the following disclaimer.
+////  * Redistributions in binary form must reproduce the above copyright
+////    notice, this list of conditions and the following disclaimer in the
+////    documentation and/or other materials provided with the distribution.
+////  * Neither the name of NVIDIA CORPORATION nor the names of its
+////    contributors may be used to endorse or promote products derived
+////    from this software without specific prior written permission.
+////
+//// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS ''AS IS'' AND ANY
+//// EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
+//// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
+//// PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE COPYRIGHT OWNER OR
+//// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
+//// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
+//// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
+//// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
+//// OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+//// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
+//// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+////
+//// Copyright (c) 2008-2022 NVIDIA Corporation. All rights reserved.
+//// Copyright (c) 2004-2008 AGEIA Technologies, Inc. All rights reserved.
+//// Copyright (c) 2001-2004 NovodeX AG. All rights reserved.  
 //
-// THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS ''AS IS'' AND ANY
-// EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
-// IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR
-// PURPOSE ARE DISCLAIMED.  IN NO EVENT SHALL THE COPYRIGHT OWNER OR
-// CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL,
-// EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED TO,
-// PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR
-// PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
-// OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
-// (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
-// OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+//#define RENDER_SNIPPET
 //
-// Copyright (c) 2008-2022 NVIDIA Corporation. All rights reserved.
-// Copyright (c) 2004-2008 AGEIA Technologies, Inc. All rights reserved.
-// Copyright (c) 2001-2004 NovodeX AG. All rights reserved.  
-
-#define RENDER_SNIPPET
-
-#ifdef RENDER_SNIPPET
-
-#include <vector>
-
-#include "PxPhysicsAPI.h"
-
-#include "../snippetrender/SnippetRender.h"
-#include "../snippetrender/SnippetCamera.h"
-
-using namespace physx;
-
-extern bool initPhysics();
-extern void stepPhysics();	
-extern void cleanupPhysics();
-
-extern PxScene* gScene;
-
-namespace
-{
-Snippets::Camera* sCamera;
-
-void renderCallback()
-{
-	stepPhysics();
-
-	Snippets::startRender(sCamera);
-
-	PxU32 nbActors = gScene->getNbActors(PxActorTypeFlag::eRIGID_DYNAMIC | PxActorTypeFlag::eRIGID_STATIC);
-	if(nbActors)
-	{
-		const PxVec3 dynColor(1.0f, 0.5f, 0.25f);
-
-		std::vector<PxRigidActor*> actors(nbActors);
-		gScene->getActors(PxActorTypeFlag::eRIGID_DYNAMIC | PxActorTypeFlag::eRIGID_STATIC, reinterpret_cast<PxActor**>(&actors[0]), nbActors);
-		Snippets::renderActors(&actors[0], static_cast<PxU32>(actors.size()), true, dynColor);
-	}
-
-	Snippets::finishRender();
-}
-
-void exitCallback(void)
-{
-	delete sCamera;
-	cleanupPhysics();
-}
-}
-
-void renderLoop()
-{
-	sCamera = new Snippets::Camera(PxVec3(10.0f, 10.0f, 10.0f), PxVec3(-0.6f,-0.2f,-0.7f));
-
-	Snippets::setupDefault("PhysX Snippet Vehicle2", sCamera, NULL, renderCallback, exitCallback);
-
-	if (initPhysics())
-	{
-		//glutMainLoop();
-	}
-}
-
-#endif
+//#ifdef RENDER_SNIPPET
+//
+//#include <vector>
+//
+//#include "PxPhysicsAPI.h"
+//
+//#include "../snippetrender/SnippetRender.h"
+//#include "../snippetrender/SnippetCamera.h"
+//
+//using namespace physx;
+//
+//extern bool initPhysics();
+//extern void stepPhysics();	
+//extern void cleanupPhysics();
+//
+//extern PxScene* gScene;
+//
+//namespace
+//{
+//Snippets::Camera* sCamera;
+//
+//void renderCallback()
+//{
+//	stepPhysics();
+//
+//	Snippets::startRender(sCamera);
+//
+//	PxU32 nbActors = gScene->getNbActors(PxActorTypeFlag::eRIGID_DYNAMIC | PxActorTypeFlag::eRIGID_STATIC);
+//	if(nbActors)
+//	{
+//		const PxVec3 dynColor(1.0f, 0.5f, 0.25f);
+//
+//		std::vector<PxRigidActor*> actors(nbActors);
+//		gScene->getActors(PxActorTypeFlag::eRIGID_DYNAMIC | PxActorTypeFlag::eRIGID_STATIC, reinterpret_cast<PxActor**>(&actors[0]), nbActors);
+//		Snippets::renderActors(&actors[0], static_cast<PxU32>(actors.size()), true, dynColor);
+//	}
+//
+//	Snippets::finishRender();
+//}
+//
+//void exitCallback(void)
+//{
+//	delete sCamera;
+//	cleanupPhysics();
+//}
+//}
+//
+//void renderLoop()
+//{
+//	sCamera = new Snippets::Camera(PxVec3(10.0f, 10.0f, 10.0f), PxVec3(-0.6f,-0.2f,-0.7f));
+//
+//	Snippets::setupDefault("PhysX Snippet Vehicle2", sCamera, NULL, renderCallback, exitCallback);
+//
+//	if (initPhysics())
+//	{
+//		//glutMainLoop();
+//	}
+//}
+//
+//#endif
